+++ dart/geobase/CHANGELOG.md
@@ -11,6 +11,7 @@ NOTE: Version 1.0.0 currently under development (1.0.0-dev.0).
 🧩 Features:
 * [Handle the special case of bbox or geometries spanning the anti-meridian (longitude 180 degrees) #185](https://github.com/navibyte/geospatial/issues/185)
 * [Merge and split on Box #211](https://github.com/navibyte/geospatial/issues/211)
+* [PositionScheme for generic, geographic and projected position data #214](https://github.com/navibyte/geospatial/issues/214)
 
 🛠 Refactoring:
 * [Normalize longitude with inclusive limit on east 180 degrees #210](https://github.com/navibyte/geospatial/issues/210)